app/sensor/pi/requirements.txt
boto3
click
rpi.gpio
